# feat(golang-rewrite): rename module and move `main` pacakge to `cmd/asdf`