--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -30,7 +30,7 @@ jobs:
         with:
           go-version: '1.21.5'
       - name: Install dependencies
-        run: go get .
+        run: go get ./...
       - name: Install gofumpt for formatting
         run: go install mvdan.cc/gofumpt@latest
       - name: Run 'gofumpt'
@@ -42,7 +42,7 @@ jobs:
       - name: Run 'revive'
         run: revive -set_exit_status ./...
       - name: Vet
-        run: go vet
+        run: go vet ./...
       - name: Install staticcheck for linting
         run: go install honnef.co/go/tools/cmd/staticcheck@latest
       - name: Lint