--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -50,7 +50,7 @@ jobs:
           go-version: '1.21.5'
       - run: scripts/install_dependencies.bash
       - name: Install dependencies
-        run: go get .
+        run: go get ./...
       - name: Run Go tests
         run: go test -coverprofile=/tmp/coverage.out  -bench= -race ./...