--- a/cmd/asdf/main.go
+++ b/cmd/asdf/main.go
@@ -1,14 +1,12 @@
 // Main entrypoint for the CLI app
 package main
 
-import (
-	"asdf/cmd"
-)
+import "github.com/asdf-vm/asdf/cli"
 
 // Replaced with the real version during a typical build
 var version = "v-dev"
 
 // Placeholder for the real code
 func main() {
-	cmd.Execute(version)
+	cli.Execute(version)
 }