--- a/internal/execenv/execenv.go
+++ b/internal/execenv/execenv.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"strings"
 
-	"asdf/internal/execute"
-	"asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/execute"
+	"github.com/asdf-vm/asdf/internal/plugins"
 )
 
 const execEnvCallbackName = "exec-env"