--- a/internal/execenv/execenv_test.go
+++ b/internal/execenv/execenv_test.go
@@ -3,10 +3,9 @@ package execenv
 import (
 	"testing"
 
-	"asdf/internal/config"
-	"asdf/internal/plugins"
-	"asdf/repotest"
-
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/repotest"
 	"github.com/stretchr/testify/assert"
 )