--- a/internal/hook/hook_test.go
+++ b/internal/hook/hook_test.go
@@ -4,8 +4,7 @@ import (
 	"os/exec"
 	"testing"
 
-	"asdf/internal/config"
-
+	"github.com/asdf-vm/asdf/internal/config"
 	"github.com/stretchr/testify/assert"
 )