--- a/internal/info/info.go
+++ b/internal/info/info.go
@@ -7,9 +7,9 @@ import (
 	"os"
 	"text/tabwriter"
 
-	"asdf/internal/config"
-	"asdf/internal/execute"
-	"asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/execute"
+	"github.com/asdf-vm/asdf/internal/plugins"
 )
 
 // Print info output to STDOUT