--- a/internal/installs/installs.go
+++ b/internal/installs/installs.go
@@ -8,10 +8,10 @@ import (
 	"os"
 	"path/filepath"
 
-	"asdf/internal/config"
-	"asdf/internal/data"
-	"asdf/internal/plugins"
-	"asdf/internal/toolversions"
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/data"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/toolversions"
 )
 
 // Installed returns a slice of all installed versions for a given plugin