--- a/internal/installs/installs_test.go
+++ b/internal/installs/installs_test.go
@@ -5,12 +5,11 @@ import (
 	"path/filepath"
 	"testing"
 
-	"asdf/internal/config"
-	"asdf/internal/installtest"
-	"asdf/internal/plugins"
-	"asdf/internal/toolversions"
-	"asdf/repotest"
-
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/installtest"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/toolversions"
+	"github.com/asdf-vm/asdf/repotest"
 	"github.com/stretchr/testify/assert"
 )