--- a/internal/resolve/resolve.go
+++ b/internal/resolve/resolve.go
@@ -9,9 +9,9 @@ import (
 	"path"
 	"strings"
 
-	"asdf/internal/config"
-	"asdf/internal/plugins"
-	"asdf/internal/toolversions"
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/toolversions"
 )
 
 // ToolVersions represents a tool along with versions specified for it