--- a/internal/resolve/resolve_test.go
+++ b/internal/resolve/resolve_test.go
@@ -6,10 +6,9 @@ import (
 	"path/filepath"
 	"testing"
 
-	"asdf/internal/config"
-	"asdf/internal/plugins"
-	"asdf/repotest"
-
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/repotest"
 	"github.com/stretchr/testify/assert"
 )