--- a/internal/shims/shims.go
+++ b/internal/shims/shims.go
@@ -11,14 +11,13 @@ import (
 	"slices"
 	"strings"
 
-	"asdf/internal/config"
-	"asdf/internal/hook"
-	"asdf/internal/installs"
-	"asdf/internal/paths"
-	"asdf/internal/plugins"
-	"asdf/internal/resolve"
-	"asdf/internal/toolversions"
-
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/hook"
+	"github.com/asdf-vm/asdf/internal/installs"
+	"github.com/asdf-vm/asdf/internal/paths"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/resolve"
+	"github.com/asdf-vm/asdf/internal/toolversions"
 	"golang.org/x/sys/unix"
 )