--- a/internal/shims/shims_test.go
+++ b/internal/shims/shims_test.go
@@ -8,13 +8,12 @@ import (
 	"strings"
 	"testing"
 
-	"asdf/internal/config"
-	"asdf/internal/installs"
-	"asdf/internal/installtest"
-	"asdf/internal/plugins"
-	"asdf/internal/toolversions"
-	"asdf/repotest"
-
+	"github.com/asdf-vm/asdf/internal/config"
+	"github.com/asdf-vm/asdf/internal/installs"
+	"github.com/asdf-vm/asdf/internal/installtest"
+	"github.com/asdf-vm/asdf/internal/plugins"
+	"github.com/asdf-vm/asdf/internal/toolversions"
+	"github.com/asdf-vm/asdf/repotest"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/sys/unix"
 )